requirements.txt
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gymnasium==1.1.1
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
ta==0.11.0
tomli==2.2.1
torch==2.6.0
typing_extensions==4.13.1
tzdata==2025.2
